Get rid of circular import for get_last_checkpoint_epoch
