+++ .gitignore
@@ -10,3 +10,4 @@ corebehrt/sandbox
 /profiling
 /tmp
 corebehrt/main/create_data_profiling.py
+**tmp**
