+++ corebehrt/common/loader.py
@@ -8,7 +8,7 @@ from transformers import ModernBertConfig
 
 from corebehrt.common.config import Config, load_config
 from corebehrt.modules.dataset import PatientDataset
-from corebehrt.modules.trainer.setup import get_last_checkpoint_epoch
+from corebehrt.functional.setup import get_last_checkpoint_epoch
 
 logger = logging.getLogger(__name__)  # Get the logger for this module
