+++ corebehrt/modules/trainer/setup.py
@@ -1,6 +1,5 @@
 import logging
 import os
-import re
 from os.path import join
 from typing import List, Optional, Tuple
 
@@ -13,6 +12,7 @@ from transformers import ModernBertConfig
 from corebehrt.common.config import Config, instantiate_class
 from corebehrt.common.loader import ModelLoader, load_model_cfg_from_checkpoint
 from corebehrt.common.setup import CHECKPOINTS_DIR
+from corebehrt.functional.setup import get_last_checkpoint_epoch
 from corebehrt.modules.model.model import BertEHRModel, BertForFineTuning
 
 logger = logging.getLogger(__name__)
@@ -229,21 +229,3 @@ def get_sampler(cfg, outcomes: List[int]):
         return sampler
     else:
         return None
-
-
-def get_last_checkpoint_epoch(checkpoint_folder: str) -> int:
-    """Returns the epoch of the last checkpoint."""
-    # Regular expression to match the pattern retry_XXX
-    pattern = re.compile(r"checkpoint_epoch(\d+)_end\.pt$")
-    max_epoch = None
-    for filename in os.listdir(checkpoint_folder):
-        match = pattern.match(filename)
-        if match:
-            epoch = int(match.group(1))
-            if max_epoch is None or epoch > max_epoch:
-                max_epoch = epoch
-    # Return the folder with the maximum retry number
-    if max_epoch is None:
-        raise ValueError("No checkpoint found in folder {}".format(checkpoint_folder))
-
-    return max_epoch
